+++ packages/babel-plugin-transform-lwc-class/src/__tests__/wire-decorator.spec.js
@@ -1,5 +1,39 @@
 const pluginTest = require('./utils/test-transform').pluginTest(require('../index'));
 
+const wireMetadataParameterTest =
+    (testName, {declaration = '', wireParameters = '', expectedStaticParameters = {}, expectedVariableParameters = {}}) => {
+        pluginTest(
+            testName,
+            `
+        import { wire } from 'lwc';
+        import { getRecord } from 'recordDataService';
+        ${declaration};
+        export default class Test {
+            @wire(getRecord, { ${wireParameters.join(',')} })
+            recordData;
+        }
+    `,
+            {
+                output: {
+                    metadata: {
+                        decorators: [{
+                            type: 'wire',
+                            targets: [
+                                {
+                                    adapter: { name: 'getRecord', reference: 'recordDataService' },
+                                    name: 'recordData',
+                                    params: expectedVariableParameters,
+                                    static: expectedStaticParameters ,
+                                    type: 'property',
+                                }
+                            ],
+                        }]
+                    },
+                },
+            },
+        );
+};
+
 describe('Transform property', () => {
     pluginTest('transforms wired field', `
         import { wire } from 'lwc';
@@ -126,7 +160,7 @@ Test.wire = {
   }
 };`
         }
-    })
+    });
 
     pluginTest('decorator accepts an optional config object as second parameter', `
         import { wire } from 'lwc';
@@ -331,7 +365,7 @@ Test.wire = {
 
 describe('Metadata', () => {
     pluginTest(
-        'gather track metadata',
+        'gather wire metadata',
         `
         import { wire } from 'lwc';
         import { getFoo } from 'data-service';
@@ -352,14 +386,14 @@ describe('Metadata', () => {
                             adapter: { name: 'getFoo', reference: 'data-service' },
                             name: 'wiredProp',
                             params: { key1: 'prop1' },
-                            static: { key2: ['fixed'] },
+                            static: { key2: { value: ['fixed'], type: 'array' } },
                             type: 'property',
                         },
                         {
                             adapter: { name: 'getFoo', reference: 'data-service' },
                             name: 'wiredMethod',
                             params: { key1: 'prop1' },
-                            static: { key2: ['fixed'] },
+                            static: { key2: { value: ['fixed'], type: 'array' } },
                             type: 'method',
                         }],
                     }]
@@ -367,4 +401,91 @@ describe('Metadata', () => {
             },
         },
     );
+
+    wireMetadataParameterTest('when constant initialised to a string-literal should extract the value',
+        { declaration: `const stringConstant = '123';`,
+            wireParameters: ['userId: stringConstant'],
+            expectedStaticParameters: { userId: { value: '123', type: 'string'} } });
+
+    wireMetadataParameterTest('when constant initialised to a number-literal should extract the value',
+        { declaration: `const numberConstant = 100;`,
+            wireParameters: ['size: numberConstant'],
+            expectedStaticParameters: { size: { value: 100, type: 'number'} } });
+
+    wireMetadataParameterTest('when constant initialised to a boolean-literal should extract the value',
+        { declaration: `const booleanConstant = true;`,
+            wireParameters: ['isRegistered: booleanConstant'],
+            expectedStaticParameters: { isRegistered: { value: true, type: 'boolean'} } });
+
+    wireMetadataParameterTest('when constant initialised as a reference to another should mark as unresolved',
+        { declaration: `const stringConstant = '123'; const referenceConstant = stringConstant;`,
+            wireParameters: ['recordId: referenceConstant'],
+            expectedStaticParameters: { recordId: { type: 'unresolved', value: 'identifier' } } });
+
+    wireMetadataParameterTest('when importing a default export from a module should reference the name of the module',
+        { declaration: `import id from '@salesforce/user/Id';`,
+            wireParameters: ['recordId: id'],
+            expectedStaticParameters: { recordId: { value: '@salesforce/user/Id', type: 'module' } } });
+
+    wireMetadataParameterTest('when parameter reference missing should mark as unresolved',
+        { wireParameters: ['recordId: id'],
+            expectedStaticParameters: { recordId: { type: 'unresolved', value: 'identifier'} } });
+
+    wireMetadataParameterTest('when importing named export with "as" from a module should reference the name of the module',
+        { declaration: `import { id as currentUserId } from '@salesforce/user/Id';`,
+            wireParameters: ['recordId: currentUserId'],
+            expectedStaticParameters: { recordId: { value: '@salesforce/user/Id', type: 'module' } } });
+
+    wireMetadataParameterTest('when importing a named export from a module should reference the name of the module',
+        { declaration: `import { id } from '@salesforce/user/Id';`,
+            wireParameters: ['recordId: id'],
+            expectedStaticParameters: { recordId: { value: '@salesforce/user/Id', type: 'module' } } });
+
+    wireMetadataParameterTest('when importing from a relative module should reference the name of the module',
+        { declaration: `import id from './someReference.js';`,
+            wireParameters: ['recordId: id'],
+            expectedStaticParameters: { recordId: { value: './someReference.js', type: 'module' } } });
+
+    wireMetadataParameterTest('when referencing a "let" variable should mark as unresolved',
+        { declaration: `let userId = '123';`,
+            wireParameters: ['recordId: userId'],
+            expectedStaticParameters: { recordId: { type: 'unresolved', value: 'identifier'} } });
+
+    wireMetadataParameterTest('when referencing a member expression, should mark as unresolved',
+        { declaration: `const data = {userId: '123'};`,
+            wireParameters: ['recordId: data.userId'],
+            expectedStaticParameters: { recordId: { type: 'unresolved', value: 'member_expression' } } });
+
+    wireMetadataParameterTest('when an inline string-literal initialization is used, should use value',
+        { wireParameters: ['recordId: "123"'],
+            expectedStaticParameters: { recordId: { value: '123', type: 'string' } } });
+
+    wireMetadataParameterTest('when an inline numeric-literal initialization is used, should use value',
+        { wireParameters: ['size: 100'],
+            expectedStaticParameters: { size: { value: 100, type: 'number' } } });
+
+    wireMetadataParameterTest('when an inline float-literal initialization is used, should use value',
+        { wireParameters: ['underPrice: 50.50'],
+            expectedStaticParameters: { underPrice: { value: 50.50, type: 'number' } } });
+
+    wireMetadataParameterTest('when an inline boolean-literal initialization is used, should use value',
+        { wireParameters: ['isRegistered: true'],
+            expectedStaticParameters: { isRegistered: { value: true, type: 'boolean' } } });
+
+    wireMetadataParameterTest('when $foo parameters are used, should use name of the parameter',
+        { wireParameters: ['recordId: "$userId"'],
+            expectedVariableParameters: { recordId: 'userId' } });
+
+    wireMetadataParameterTest('when $foo parameters with dots are used, should use name of the parameter',
+        { wireParameters: ['recordId: "$userRecord.Id"'],
+            expectedVariableParameters: { recordId: 'userRecord.Id' } });
+
+    wireMetadataParameterTest('when inline array with a non-string literal is used, should mark as unresolved',
+        { declaration: `const bar = '123';`,
+            wireParameters: ['fields: ["foo", bar]'],
+            expectedStaticParameters: { fields: {type: 'unresolved', value: 'array_expression'} } });
+
+    wireMetadataParameterTest('when inline array with literals is used, should have the array',
+        {wireParameters: ['data: ["foo", 123, false]'],
+            expectedStaticParameters: { data: {type: 'array', value: ['foo', 123, false]} } });
 });