+++ packages/benchmark/best.config.js
@@ -1,26 +1,7 @@
 module.exports = {
     projectName: 'lwc-engine-benchmark',
-    plugins: [
-        ['rollup-plugin-lwc-compiler', {
-            rootDir: '<rootDir>/src/',
-            mode: 'prod',
-        }],
-    ],
-    benchmarkOnClient: false,
-    benchmarkIterations: 60,
-    runnerConfig: [
-        {
-            "runner": '@best/runner-headless',
-            "name": "default",
-        },
-        {
-            "runner": '@best/runner-remote',
-            "name": "remote",
-            "config": {
-                "host": "http://best-agent-pool.lwcjs.org",
-                "options": { path: '/best' },
-                "remoteRunner": "@best/runner-headless"
-            }
-        }
+    projects: [
+        '<rootDir>/best.ie11.config.js',
+        '<rootDir>/best.headless.config.js'
     ],
 };