+++ packages/lwc-compiler/src/__tests__/fixtures/namespaced-compilation/utils.js
@@ -1,3 +1,0 @@
-export function method(name) {
-    return 'utilities was called with the name: ' + name;
-}