+++ packages/lwc-compiler/src/compiler/options.ts
@@ -2,7 +2,6 @@ import { isBoolean, isString, isUndefined, isObject } from "../utils";
 
 const DEFAULT_OPTIONS = {
     baseDir: "",
-    namespaceMapping: {},
 };
 
 const DEFAULT_STYLESHEET_CONFIG: NormalizedStylesheetConfig = {
@@ -49,14 +48,9 @@ export interface BundleFiles {
     [filename: string]: string;
 }
 
-export interface NamespaceMapping {
-    [name: string]: string;
-}
-
 export interface CompilerOptions {
     name: string;
     namespace: string;
-    namespaceMapping?: NamespaceMapping;
     files: BundleFiles;
     /**
      * An optional directory prefix that contains the specified components
@@ -69,7 +63,6 @@ export interface CompilerOptions {
 
 export interface NormalizedCompilerOptions extends CompilerOptions {
     outputConfig: NormalizedOutputConfig;
-    namespaceMapping: NamespaceMapping;
     stylesheetConfig: NormalizedStylesheetConfig;
 }
 
@@ -121,28 +114,6 @@ export function validateOptions(options: CompilerOptions) {
             throw new TypeError(
                 `Unexpected file content for "${key}". Expected a string, received "${value}".`
             );
-        }
-    }
-
-    if (options.namespaceMapping) {
-        if (!isObject(options.namespaceMapping)) {
-            throw new TypeError(
-                `Expected an object for namespaceMapping, received "${options.namespaceMapping}"`,
-            );
-        }
-
-        for (const [key, value] of Object.entries(options.namespaceMapping)) {
-            if (!isString(key)) {
-                throw new TypeError(
-                    `Expected a string key in namespaceMapping, received "${key}"`,
-                );
-            }
-
-            if (!isString(value)) {
-                throw new TypeError(
-                    `Expected a string value in namespaceMapping.${key}, received "${value}"`,
-                );
-            }
         }
     }