+++ packages/lwc-engine/src/framework/restrictions.ts
@@ -255,6 +255,15 @@ function getComponentRestrictionsDescriptors(cmp: ComponentInterface): PropertyD
             configurable: true,
             writable: true,
         },
+        tagName: {
+            get(this: ComponentInterface) {
+                throw new Error(
+                    `Usage of property \`tagName\` is disallowed because the component itself does not know which tagName will be used to create the element, therefore writing code that check for that value is error prone.`,
+                );
+            },
+            enumerable: true,
+            configurable: true,
+        },
     };
 }