+++ packages/lwc-integration/scripts/build.js
@@ -4,8 +4,8 @@ const babel = require('@babel/core');
 const rollup = require('rollup');
 const templates = require('../src/shared/templates.js');
 const rollupLwcCompilerPlugin = require('rollup-plugin-lwc-compiler');
-const rollupCompatPlugin = require('rollup-plugin-compat').default;
-const babelPresetCompat = require('babel-preset-compat');
+const rollupCompatPlugin = require('rollup-plugin-compat');
+const rollupReplacePlugin = require('rollup-plugin-replace');
 const compatPolyfills = require('compat-polyfills');
 
 // -- Build Config -------------------------------------------
@@ -89,15 +89,11 @@ const baseInputConfig = {
     plugins: [
         entryPointResolverPlugin(),
         rollupLwcCompilerPlugin({
-            mode,
             exclude: `**/*${testSufix}`,
-            resolveFromPackages: false,
-            compat: {
-                // In order to build faster we manually add compat artifacts later
-                downgrade: false,
-                polyfills: false
-            }
+            resolveFromPackages: false
         }),
+        isCompat && rollupCompatPlugin({ polyfills: false }),
+        isProd && rollupReplacePlugin({ 'process.env.NODE_ENV': JSON.stringify('production') }),
         testCaseComponentResolverPlugin(),
     ].filter(Boolean)
 };