+++ packages/lwc-jest-transformer/src/test/modules/example/resourceUrl/resourceUrl.js
@@ -1,8 +1,8 @@
 import { LightningElement } from 'lwc';
-import mockedImport from '@salesforce/resource-url/c.mocked';
-import unmockedImport from '@salesforce/resource-url/c.unmocked';
+import mockedImport from '@salesforce/resourceUrl/c.mocked';
+import unmockedImport from '@salesforce/resourceUrl/c.unmocked';
 
-export default class Labels extends LightningElement {
+export default class ResourceUrl extends LightningElement {
     mockedResource = mockedImport;
     unmockedResource = unmockedImport;
 }