+++ packages/rollup-plugin-lwc-compiler/src/index.js
@@ -1,54 +1,16 @@
 const fs = require("fs");
 const path = require("path");
-const babel = require("@babel/core");
-const minify = require("babel-preset-minify");
 const compiler = require("lwc-compiler");
 const pluginUtils = require("rollup-pluginutils");
-const replacePlugin = require("rollup-plugin-replace");
 const lwcResolver = require("lwc-module-resolver");
-const rollupCompatPlugin = require("rollup-plugin-compat").default;
+const { getModuleQualifiedName, getLwcEnginePath } = require('./utils');
+const { DEFAULT_OPTIONS, DEFAULT_MODE } = require("./constants");
 
-
-const { DEFAULT_NS, DEFAULT_OPTIONS, DEFAULT_MODE } = require("./constants");
-
-function getModuleQualifiedName(file) {
-    const registry = {
-        entry: file,
-        moduleSpecifier: null,
-        moduleName: null,
-        moduleNamespace: DEFAULT_NS
-    };
-
-    const rootParts = path.dirname(file).split(path.sep);
-
-    registry.moduleName = rootParts.pop();
-    registry.moduleNamespace = rootParts.pop();
-
-    return registry;
-}
-
-function normalizeResult(result) {
-    return { code: result.code || result, map: result.map || { mappings: "" } };
-}
-
-/*
-    API for rollup-compat plugin:
-    {
-        disableProxyTransform?: boolean,
-        onlyProxyTransform?: boolean,
-        downgrade?: boolean,
-        resolveProxyCompat?
-        polyfills?
-    }
- */
 module.exports = function rollupLwcCompiler(pluginOptions = {}) {
     const { include, exclude } = pluginOptions;
     const filter = pluginUtils.createFilter(include, exclude);
-    const mergedPluginOptions = Object.assign({}, DEFAULT_OPTIONS, pluginOptions,);
-    const { mode, compat } = mergedPluginOptions;
-
-    // We will compose compat plugin on top of this one
-    const rollupCompatInstance = rollupCompatPlugin(compat);
+    const mergedPluginOptions = Object.assign({}, DEFAULT_OPTIONS, pluginOptions);
+    const { resolveFromPackages, resolveFromSource } = mergedPluginOptions;
 
     // Closure to store the resolved modules
     let modulePaths = {};
@@ -60,12 +22,8 @@ module.exports = function rollupLwcCompiler(pluginOptions = {}) {
             modulePaths = {};
             const entry = rollupOptions.input || rollupOptions.entry;
             const entryDir = mergedPluginOptions.rootDir || path.dirname(entry);
-            const externalPaths = mergedPluginOptions.resolveFromPackages
-                ? lwcResolver.resolveLwcNpmModules(mergedPluginOptions)
-                : {};
-            const sourcePaths = mergedPluginOptions.resolveFromSource
-                ? lwcResolver.resolveModulesInDir(entryDir, mergedPluginOptions)
-                : {};
+            const externalPaths = resolveFromPackages ? lwcResolver.resolveLwcNpmModules(mergedPluginOptions) : {};
+            const sourcePaths = resolveFromSource ? lwcResolver.resolveModulesInDir(entryDir, mergedPluginOptions) : {};
             Object.assign(modulePaths, externalPaths, sourcePaths);
         },
 
@@ -77,15 +35,9 @@ module.exports = function rollupLwcCompiler(pluginOptions = {}) {
 
             // Normalize relative import to absolute import
             if (importee.startsWith(".") && importer) {
-                const normalizedPath = path.resolve(
-                    path.dirname(importer),
-                    importee
-                );
+                const normalizedPath = path.resolve(path.dirname(importer), importee);
                 return pluginUtils.addExtension(normalizedPath);
             }
-
-            // Check if compat needs to resolve this file
-            return rollupCompatInstance.resolveId(importee, importer);
         },
 
         load(id) {
@@ -95,12 +47,10 @@ module.exports = function rollupLwcCompiler(pluginOptions = {}) {
             if (!exists && isCSS) {
                 return "";
             }
-
-            // Check if compat knows how to load this file
-            return rollupCompatInstance.load(id);
         },
 
-        async transform(code, id) {
+        async transform(src, id) {
+            // Filter user-land config and lwc import
             if (!filter(id)) {
                 return;
             }
@@ -109,57 +59,14 @@ module.exports = function rollupLwcCompiler(pluginOptions = {}) {
             const moduleEntry = Object.values(modulePaths).find(r => id === r.entry);
             const moduleRegistry = moduleEntry || getModuleQualifiedName(id, mergedPluginOptions);
 
-            let result = code;
+            const { code, map } = await compiler.transform(src, id, {
+                mode: DEFAULT_MODE, // Use always default mode since any other (prod or compat) will be resolved later
+                name: moduleRegistry.moduleName,
+                namespace: moduleRegistry.moduleNamespace,
+                moduleSpecifier: moduleRegistry.moduleSpecifier
+            });
 
-            if (!rollupCompatInstance.knownCompatModule(id)) {
-                result = await compiler.transform(code, id, {
-                    mode: DEFAULT_MODE, // Use always default mode since any other (prod or compat) will be resolved later
-                    name: moduleRegistry.moduleName,
-                    namespace: moduleRegistry.moduleNamespace,
-                    moduleSpecifier: moduleRegistry.moduleSpecifier
-                });
-            }
-
-            result = normalizeResult(result);
-
-            if (mode === "compat" || mode === "prod_compat") {
-                result = normalizeResult(
-                    rollupCompatInstance.transform(result.code, id)
-                );
-            }
-
-            return { code: result.code, map: result.map };
-        },
-
-        transformBundle(code) {
-            if (mode === "compat" || mode === "prod_compat") {
-                code = rollupCompatInstance.transformBundle(code);
-            }
-            let result = undefined;
-            if (mode === "prod" || mode === "prod_compat") {
-                const rollupReplace = replacePlugin({
-                    "process.env.NODE_ENV": JSON.stringify("production")
-                });
-                const resultReplace = rollupReplace.transform(
-                    code,
-                    "$__tmpBundleSrc"
-                );
-
-                const transformConfig = {
-                    babelrc: false,
-                    sourceMaps: true,
-                    parserOpts: { plugins: ["*"] },
-                    presets: [[minify, { guards: false, evaluate: false }]],
-                };
-
-                const output = babel.transform(
-                    resultReplace ? resultReplace.code : code,
-                    transformConfig
-                );
-
-                result = output.code;
-            }
-            return result || code;
+            return { code, map };
         }
     };
 };